Creation form fixes
Fixed /createUser redirecting to /auth when the page should be accessible for unauthorized users.

Added alerts and minor error-handling to the fetch functions of creation components, notifying the user about the successful or unsuccessful creation of an account/case.

Added an extra check for ownerFullName as a mandatory createCase argument.